refactor: move service inside server directory
