+++ webapp/src/app/[locale]/session/page.tsx
@@ -7,7 +7,7 @@ import { useEffect, useState } from 'react';
 import SessionClient from '@/client/service/session/implement';
 import { RecommendationProvider } from '@/components/recommendation/RecommendationContext';
 import RecommendationResult from '@/components/recommendation/RecommendationResult';
-import type Session from '@/service/database/session/model';
+import type Session from '@/server/service/database/session/model';
 
 const SessionResult = () => {
   const searchParams = useSearchParams();
