+++ webapp/src/server/service/cache/interface.ts
