+++ webapp/src/server/service/map/place/interface.ts
