+++ webapp/src/server/service/map/place/restaurant/interface.ts
