+++ webapp/src/server/service/s3/interface.ts
